% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurin_helpers.R
\name{aurin_api_credentials}
\alias{aurin_api_credentials}
\title{Add a username and password to an existing wfs_connection xml file.}
\usage{
aurin_api_credentials(username, password)
}
\arguments{
\item{username}{AURIN API username as provided by AURIN.}

\item{password}{AURIN API password as provided by AURIN.}
}
\description{
Edits an existing aurin_wfs_connection.xml file to
add a username and password. This function is not exported and should not
be called directly.
}
