% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurin_helpers.R
\name{aurin_authorise_api}
\alias{aurin_authorise_api}
\title{Authorise project to use AURIN API.}
\usage{
aurin_authorise_api(username = NULL, password = NULL)
}
\arguments{
\item{username}{AURIN API username as provided by AURIN.}

\item{password}{AURIN API password as provided by AURIN.}
}
\description{
This creates an aurin_wfs_connection.xml file if one
does not already exist. If you specify a username and password, it will
also add them to the .xml file. See \verb{https://aurin.org.au/resources/aurin-apis/}
to register for access to the API.
}
\examples{
\dontrun{aurin_authorise_api()}
}
