% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurin_helpers.R
\name{aurin_create_wfs_connection}
\alias{aurin_create_wfs_connection}
\title{Create a aurin_wfs_connection.xml file for use with the AURIN API.}
\usage{
aurin_create_wfs_connection()
}
\description{
Create a aurin_wfs_connection.xml file for use with the AURIN API.
}
