% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurin_helpers.R
\name{aurin_download_file}
\alias{aurin_download_file}
\title{Download a file using the AURIN API.}
\usage{
aurin_download_file(api_id, out_file_name, out_folder = "out", ...)
}
\arguments{
\item{api_id}{The API ID of the dataset you wish to download. See \code{?aurin_id} for methods of getting the API ID.}

\item{out_file_name}{What name to save the downloaded file.}

\item{out_folder}{A path to where the file should be downloaded. The default is a folder \code{out} inside the current working directory.}

\item{...}{additional options passed on to ogr2ogr2. See \code{?ogr2ogr2} for available options.}
}
\value{
The path of the downloaded file
}
\description{
Download a file using the AURIN API.
}
\examples{
\dontrun{aurin_download_file(api_id = aurin_id("DSS - National Public Toilets (Point) 2017"),
out_file_name = "toilets")}
}
