% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurin_helpers.R
\name{aurin_id}
\alias{aurin_id}
\title{Fetch the API ID for an AURIN dataset.}
\usage{
aurin_id(search, exact = FALSE)
}
\arguments{
\item{search}{(string) Search term.}

\item{exact}{(logical) If you know the exact name of the dataset you are looking for, specify exact = TRUE.
This should return only one API ID.}
}
\value{
AURIN API ID for a given dataset
}
\description{
Fetch the API ID for an AURIN dataset.
}
\examples{
\dontrun{aurin_api("DSS - National Public Toilets (Point) 2017", exact = TRUE)}

}
