% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurin_helpers.R
\name{read_aurin}
\alias{read_aurin}
\title{Read a downloaded AURIN Dataset into R memory.}
\usage{
read_aurin(aurin_file)
}
\arguments{
\item{aurin_file}{A path to the file downloaded from the AURIN API.}
}
\value{
A simple features file
}
\description{
Read a downloaded AURIN Dataset into R memory.
}
\examples{
\dontrun{read_aurin("out/aurin_file.geoJSON")}
}
